backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
anthropic==0.25.0
sentence-transformers==2.7.0
chromadb==0.5.0
mysqlclient==2.2.4
python-dotenv==1.0.1
selenium==4.20.0
requests==2.31.0
